requirements.txt
Flask==2.3.2
Flask-Session==0.5.0
pymemcache==4.0.0
PyYAML==6.0.1
slackclient==2.9.4